mqtt/fixtures_requests/app/requirements.txt
fastapi
fastapi[standard]
psycopg2-binary
sqlalchemy
uvicorn
requests
paho-mqtt